refactor: dispatch a single action on registration client :ok_hand:
