+++ mobile/test/reducers/settings.spec.js
@@ -1,6 +1,6 @@
 import reducer, { initialState } from '../../src/reducers/settings'
-import { INIT_STATE } from '../../../src/actions'
-import { SET_URL, ERROR, SET_CLIENT, SET_ANALYTICS } from '../../src/actions/settings'
+import { INIT_STATE } from '../../../src/actions/settings'
+import { SET_URL, ERROR, SET_ANALYTICS } from '../../src/actions/settings'
 
 describe('settings reducers', () => {
   it('should return the initial state', () => {
@@ -28,10 +28,5 @@ describe('settings reducers', () => {
   it('should handle INIT_STATE', () => {
     expect(reducer({serverUrl: 'serverUrl', backupImages: true, error: 'error'}, {type: INIT_STATE}))
     .toEqual(initialState)
-  })
-
-  it('should set a client into the state with "SET_CLIENT"', () => {
-    const client = { someParameter: 'Some Value' }
-    expect(reducer(undefined, { type: SET_CLIENT, client }).client).toEqual(client)
   })
 })
