+++ src/actions/index.js
@@ -4,7 +4,6 @@ import { openWithOfflineError, openWithNoAppError } from '../../mobile/src/actio
 
 import { ROOT_DIR_ID, TRASH_DIR_ID } from '../constants/config.js'
 
-export const INIT_STATE = 'INIT_STATE'
 export const LOCATION_CHANGE = 'LOCATION_CHANGE'
 export const OPEN_FOLDER = 'OPEN_FOLDER'
 export const OPEN_FOLDER_SUCCESS = 'OPEN_FOLDER_SUCCESS'
@@ -34,8 +33,6 @@ export const DOWNLOAD_FILE_E_OFFLINE = 'DOWNLOAD_FILE_E_OFFLINE'
 export const OPEN_FILE_WITH = 'OPEN_FILE_WITH'
 export const OPEN_FILE_E_OFFLINE = 'OPEN_FILE_E_OFFLINE'
 export const OPEN_FILE_E_NO_APP = 'OPEN_FILE_E_NO_APP'
-export const SET_OFFLINE = 'SET_OFFLINE'
-export const SET_FIRST_REPLICATION = 'SET_FIRST_REPLICATION'
 
 const extractFileAttributes = f => Object.assign({}, f.attributes, { id: f._id })
 const toServer = f => Object.assign({}, { attributes: f }, { _id: f.id })
@@ -318,15 +315,3 @@ export const actionMenuLoaded = (menu) => ({
   type: 'HIDE_SPINNER',
   menu
 })
-
-export const setOffline = (offline) => ({
-  type: SET_OFFLINE,
-  offline
-})
-
-export const setFirstReplication = (firstReplication) => ({
-  type: SET_FIRST_REPLICATION,
-  firstReplication
-})
-
-export const initializeState = () => ({ type: INIT_STATE })
